requirements.txt
pillow
tk
flask
requests
numpy
torch
torchvision
tensorflow
keras
click
future
boto3
python-dotenv
